fix build for demo
- kludge: sort based on "Importance"
- kludge: disk reload onResume (TaskListFragment)
- EditTaskDialog needs work

+++ TaskList/app/src/main/java/com/github/willpascucci/tasklist/model/TaskList.java
@@ -40,7 +40,7 @@ public class TaskList {
         ActiveAndroid.endTransaction();
     }
 
-    public static void updatePriority(int id) {
+    public static void updatePriority(long id) {
         ActiveAndroid.beginTransaction();
         Task t = Task.get(id);
         setPriority(t);                                             //edit by Sahil
